Moved Cart Functions
Moved all cart functions to controller/cartFunctions. cart and setCart now have to be passed into the functions as parameters
